add UnmarshalTCPOptionSlice() function
This moves the unmarshaling of the TCP Header Option field to its own exported function. This function is used to convert a byte slice in to a TCPOptionSlice.

The unmarshaling function for the TCPHeader was updated to use this exported function instead of its own implemented (where the public function originated from anyhow).