make error passing more idiomatic
This makes the errors we use in the `packets` library more idiomatic. Errors that took variables have been modified so that the variables themselves are given to the type instead of a prebuilt string. This allows consumers to pull information from the error instead of doing some sort of regex match on the string.

For error types that took no variable data, they've been converted to variables on the `packets` package using `errors.New()`. The actual error messages never changed, so it made no sense to make these their own type.

Lastly, some specs were added for...

* `ErrChecksumInvalidKind`
* `ErrUDPPayloadTooLarge`