move error-related code to packetserr sub-package
To assist with making wth GoDoc documentation for the packet a bit more readable, I've opted ot move the errors out to the `packetserr` package. Also, because of the new name of the package the `Err` suffix has been removed from all of them.

Some of the documentation has been updated as well.